# Reduce locked region in  filecache.AddFile