This improves AddFile performance by about 5%.

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                      │ /tmp/nopathinvalue │       /tmp/nopathinvalue-rl        │
                      │       sec/op       │   sec/op     vs base               │
ContainsAdd/1K/10G-24          619.8µ ± 2%   616.4µ ± 2%       ~ (p=0.199 n=50)
ContainsAdd/1K/1M-24           14.51m ± 2%   13.72m ± 1%  -5.46% (p=0.000 n=50)
geomean                        2.999m        2.908m       -3.04%

                      │ /tmp/nopathinvalue │        /tmp/nopathinvalue-rl         │
                      │   retained-heap    │ retained-heap  vs base               │
ContainsAdd/1K/10G-24         229.9k ±  3%    233.8k ±  3%       ~ (p=0.336 n=50)
ContainsAdd/1K/1M-24          65.66k ± 10%    67.69k ± 10%       ~ (p=0.809 n=50)
geomean                       122.8k          125.8k        +2.41%

                      │ /tmp/nopathinvalue │        /tmp/nopathinvalue-rl        │
                      │        B/op        │     B/op      vs base               │
ContainsAdd/1K/10G-24         165.1Ki ± 0%   165.1Ki ± 0%       ~ (p=0.139 n=50)
ContainsAdd/1K/1M-24          1.261Mi ± 0%   1.258Mi ± 0%       ~ (p=0.149 n=50)
geomean                       461.8Ki        461.1Ki       -0.14%

                      │ /tmp/nopathinvalue │       /tmp/nopathinvalue-rl        │
                      │     allocs/op      │  allocs/op   vs base               │
ContainsAdd/1K/10G-24          3.011k ± 0%   3.011k ± 0%       ~ (p=0.169 n=50)
ContainsAdd/1K/1M-24           15.42k ± 0%   15.38k ± 0%       ~ (p=0.160 n=50)
geomean                        6.813k        6.805k       -0.13%
```